[weather.ozweather] 1.2.1
